replaced enemies with zombies

+++ Dice Tower Defence/Assets/Scripts/Character/Character/CharacterStatsManager.cs
@@ -7,8 +7,8 @@ namespace UB
         public int Health;
         public int MaxHealth;
         public int Damage;
-        public int MovementSpeed;
-        public int Acceleration;
-        public int StoppingDistance;
+        public float MovementSpeed;
+        public float Acceleration;
+        public float StoppingDistance;
     }
 }
